dockerfiles/geodjango/requirements.txt
Django==1.11.7
django-leaflet==0.22.0
psycopg2==2.7.3.2
pytz==2017.3
